+++ packages/design-system/src/rwa/hooks/useFormInputs.tsx
@@ -1,4 +1,3 @@
-import { DateTimeLocalInput } from "@/connect";
 import {
   Account,
   allGroupTransactionTypes,
@@ -22,7 +21,7 @@ import {
   useEditorContext,
   useModal,
 } from "@/rwa";
-import { ReactElement, useCallback, useMemo } from "react";
+import { ReactElement, ReactNode, useCallback, useMemo } from "react";
 import {
   Control,
   FormState,
@@ -30,11 +29,12 @@ import {
   UseFormWatch,
 } from "react-hook-form";
 import { CashBalanceChange } from "../components/table/transactions/cash-balance-change";
+import { EntryTimeLabel } from "../components/inputs/entry-time-label";
 
 type Input = {
   label: string;
   Input: () => string | React.JSX.Element;
-  inputLabel?: string | null;
+  inputLabel?: ReactNode | null;
 };
 
 type Props = {
@@ -67,23 +67,21 @@ export function useFormInputs(props: Props): {
     fixedIncomes,
   } = useEditorContext();
   const { showModal } = useModal();
-
   const showCreateItemModal = useCallback(
     (tableName: TableName) => () => {
       showModal("createItem", { tableName });
     },
     [showModal],
   );
-
   return useMemo(() => {
     switch (tableName) {
       case tableNames.ASSET: {
         const tableItem = _tableItem as TableItemType<"ASSET"> | null;
         type Payload = FormInputsByTableName["ASSET"];
         const register = _register as UseFormRegister<Payload>;
-        const watch = _watch as UseFormWatch<Payload>;
         const formState = _formState as FormState<Payload>;
         const { errors } = formState;
+
         const control = _control as Control<Payload>;
         const derivedInputsToDisplay =
           operation !== "create" && !!tableItem
@@ -110,11 +108,6 @@ export function useFormInputs(props: Props): {
                 },
               ]
             : [];
-
-        const maturityInputValue =
-          operation === "view" && tableItem
-            ? tableItem.maturity
-            : watch("maturity") || tableItem?.maturity;
 
         const inputs = [
           {
@@ -189,20 +182,17 @@ export function useFormInputs(props: Props): {
           {
             label: "Maturity",
             Input: () => (
-              <DateTimeLocalInput
+              <input
+                className="h-8 w-full rounded-md bg-gray-100 px-3 disabled:bg-transparent disabled:p-0"
+                step={1}
+                type="date"
                 {...register("maturity", {
                   disabled: operation === "view",
-                  setValueAs: (value) => {
-                    if (value === "") return null;
-                    return value as string;
-                  },
                 })}
-                inputType="date"
-                name="maturity"
               />
             ),
-            inputLabel: maturityInputValue
-              ? formatDateForDisplay(maturityInputValue, true)
+            inputLabel: tableItem?.maturity
+              ? formatDateForDisplay(tableItem.maturity, true)
               : null,
           },
           {
@@ -261,8 +251,6 @@ export function useFormInputs(props: Props): {
         const { errors } = _formState as FormState<Payload>;
         const control = _control as Control<Payload>;
         const type = watch("type");
-        const entryTimeInputValue =
-          watch("entryTime") ?? new Date().toISOString();
         const isFeesTransaction = !!type && feesTransactions.includes(type);
         const isAssetTransaction =
           !!type && assetGroupTransactions.includes(type);
@@ -302,17 +290,17 @@ export function useFormInputs(props: Props): {
           {
             label: "Entry Time",
             Input: () => (
-              <DateTimeLocalInput
+              <input
+                className="h-8 w-full rounded-md bg-gray-100 px-3 disabled:bg-transparent disabled:p-0"
+                step={1}
+                type="datetime-local"
                 {...register("entryTime", {
-                  required: true,
                   disabled: operation === "view",
+                  required: "Entry time is required",
                 })}
-                name="entryTime"
               />
             ),
-            inputLabel: entryTimeInputValue
-              ? formatDateForDisplay(new Date(entryTimeInputValue))
-              : null,
+            inputLabel: <EntryTimeLabel control={control} />,
           },
           isFeesTransaction
             ? {