+++ packages/document-drive/src/server/index.ts
@@ -895,6 +895,10 @@ export class BaseDocumentDriveServer
     return documentModel;
   }
 
+  getDocumentModels() {
+    return [...this.documentModels];
+  }
+
   async addDrive(drive: DriveInput): Promise<DocumentDriveDocument> {
     const id = drive.global.id || generateUUID();
     if (!id) {
@@ -917,6 +921,8 @@ export class BaseDocumentDriveServer
     }
 
     await this._initializeDrive(id);
+
+    this.emit("driveAdded", document);
 
     return document;
   }
@@ -2308,7 +2314,6 @@ export class BaseDocumentDriveServer
     event: K,
     ...args: Parameters<DriveEvents[K]>
   ): void {
-    logger.debug(`Emitting event ${event}`, args);
     return this.emitter.emit(event, ...args);
   }
 }