+++ packages/document-model-libs/editors/types.ts
@@ -4,6 +4,20 @@ import { Action, EditorProps } from "document-model/document";
 export type EditorConfig = {
   id: string;
   disableExternalControls: boolean;
+  documentToolbarEnabled?: boolean;
+};
+
+export type EditorContextProps = {
+  readonly isAllowedToCreateDocuments: boolean;
+  readonly isAllowedToEditDocuments: boolean;
+  readonly canUndo: boolean;
+  readonly canRedo: boolean;
+  readonly undo: () => void;
+  readonly redo: () => void;
+  readonly onSwitchboardLinkClick: (() => void) | undefined;
+  readonly onExport: () => void;
+  readonly onClose: () => void;
+  readonly onShowRevisionHistory: () => void;
 };
 
 export type ExtendedEditor<