+++ packages/reactor-api/src/types.ts
@@ -1,0 +1,8 @@
+import { IDocumentDriveServer } from "document-drive";
+import { IncomingHttpHeaders } from "http";
+
+export interface Context {
+  headers: IncomingHttpHeaders;
+  driveId: string | undefined;
+  driveServer: IDocumentDriveServer;
+}