+++ packages/reactor-api/tsup.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/index.ts"],
+  splitting: true,
+  sourcemap: true,
+  clean: true,
+  format: "esm",
+  treeshake: true,
+  noExternal: ["document-drive"],
+  platform: "node",
+  target: "node20",
+  loader: {
+    ".graphql": "file",
+  },
+  dts: true,
+});