+++ packages/document-model/CHANGELOG.md
@@ -1,3 +1,107 @@
+## 2.6.0 (2024-11-06)
+
+### 🚀 Features
+
+- **document-model-libs:** regenerate lockfile ([0effee2e](https://github.com/powerhouse-inc/powerhouse/commit/0effee2e))
+- qa updates 4 ([50ac4eb5](https://github.com/powerhouse-inc/powerhouse/commit/50ac4eb5))
+
+### ❤️  Thank You
+
+- ryanwolhuter @ryanwolhuter
+
+## 2.5.0 (2024-11-06)
+
+### 🚀 Features
+
+- **document-model:** include resulting state into zip file ([#491](https://github.com/powerhouse-inc/powerhouse/pull/491))
+- **document-model-libs:** add more inputs ([37db4561](https://github.com/powerhouse-inc/powerhouse/commit/37db4561))
+- **document-model-libs:** regenerate lockfile ([cf76e91d](https://github.com/powerhouse-inc/powerhouse/commit/cf76e91d))
+- **document-model-libs:** address initial QA feedback ([7513155e](https://github.com/powerhouse-inc/powerhouse/commit/7513155e))
+
+### ❤️  Thank You
+
+- Guillermo Puente Sandoval @gpuente
+- ryanwolhuter @ryanwolhuter
+
+## 2.4.1 (2024-11-01)
+
+### 🚀 Features
+
+- **codegen:** changed bundle to esm ([24b33b50](https://github.com/powerhouse-inc/powerhouse/commit/24b33b50))
+
+### 🩹 Fixes
+
+- **ph-cli:** bundle cli as esm package and added missing dependencies ([a5a665ef](https://github.com/powerhouse-inc/powerhouse/commit/a5a665ef))
+- **codegen:** fixed bad import ([eefb7b2f](https://github.com/powerhouse-inc/powerhouse/commit/eefb7b2f))
+- bad package reference was breaking build -- also a readme update ([4121e51b](https://github.com/powerhouse-inc/powerhouse/commit/4121e51b))
+- **switchboard:** startup ([#469](https://github.com/powerhouse-inc/powerhouse/pull/469))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+- Benjamin Jordan
+- frankp.eth @froid1911
+
+## 2.4.0 (2024-10-29)
+
+### 🚀 Features
+
+- **switchboard-gui:** updated gitignore ([1fd60a3b](https://github.com/powerhouse-inc/powerhouse/commit/1fd60a3b))
+- **document-model:** fixed cjs exports ([5afce14c](https://github.com/powerhouse-inc/powerhouse/commit/5afce14c))
+- **reactor-api:** init project ([#388](https://github.com/powerhouse-inc/powerhouse/pull/388))
+- **document-model-libs:** bump deps ([68a98ee6](https://github.com/powerhouse-inc/powerhouse/commit/68a98ee6))
+- **document-model-libs:** re-add prevent default and handle operation initial schema ([ae010779](https://github.com/powerhouse-inc/powerhouse/commit/ae010779))
+- **document-model-libs:** add initial state editor ([#443](https://github.com/powerhouse-inc/powerhouse/pull/443))
+- **document-model-libs:** add json editors with sync ([76acd807](https://github.com/powerhouse-inc/powerhouse/commit/76acd807))
+- **document-model-libs:** include typedefs in initial hidden schema state ([88e184fd](https://github.com/powerhouse-inc/powerhouse/commit/88e184fd))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+- Guillermo Puente @gpuente
+- Ryan Wolhuter @ryanwolhuter
+- ryanwolhuter @ryanwolhuter
+
+## 2.3.1 (2024-10-25)
+
+### 🚀 Features
+
+- **codegen:** integrate scalars into code generation ([#436](https://github.com/powerhouse-inc/powerhouse/pull/436))
+- **document-model-libs:** added scalars as a peerDep ([#435](https://github.com/powerhouse-inc/powerhouse/pull/435))
+- **document-model-libs:** simplify editor ([1725f876](https://github.com/powerhouse-inc/powerhouse/commit/1725f876))
+- **document-model-libs:** port code from private repo ([11ef336a](https://github.com/powerhouse-inc/powerhouse/commit/11ef336a))
+- **ph-cli:** Support for dev and generate command ([#406](https://github.com/powerhouse-inc/powerhouse/pull/406))
+- **ph-cli:** added base ph-cli package setup ([#404](https://github.com/powerhouse-inc/powerhouse/pull/404))
+- **switchboard-gui:** init ([#405](https://github.com/powerhouse-inc/powerhouse/pull/405))
+- **codegen:** bundle with tsup ([3cccbdf9](https://github.com/powerhouse-inc/powerhouse/commit/3cccbdf9))
+- **document-model-libs:** update codegen dependency ([f5330139](https://github.com/powerhouse-inc/powerhouse/commit/f5330139))
+
+### 🩹 Fixes
+
+- **document-model:** fix import zip files ([#412](https://github.com/powerhouse-inc/powerhouse/pull/412))
+- **document-model-libs:** added type annotation for createDocumentStory ([6c5441f3](https://github.com/powerhouse-inc/powerhouse/commit/6c5441f3))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+- frankp.eth @froid1911
+- Guillermo Puente Sandoval @gpuente
+- ryanwolhuter @ryanwolhuter
+
+## 2.3.0 (2024-10-14)
+
+### 🚀 Features
+
+- added codegen build step ([66523278](https://github.com/powerhouse-inc/powerhouse/commit/66523278))
+- **scalars:** setup scalars project ([#370](https://github.com/powerhouse-inc/powerhouse/pull/370))
+- **design-system:** create icon components in prepare install of postinstall ([f302d65f](https://github.com/powerhouse-inc/powerhouse/commit/f302d65f))
+- **design-system:** improve bundle size and tree shaking support ([ffb68cfc](https://github.com/powerhouse-inc/powerhouse/commit/ffb68cfc))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+- Guillermo Puente Sandoval @gpuente
+
 ## 2.2.0 (2024-10-09)
 
 ### 🚀 Features