+++ packages/scalars/CHANGELOG.md
@@ -1,3 +1,152 @@
+## 1.9.0 (2024-11-11)
+
+### 🚀 Features
+
+- **ph-cli,reactor-local:** integrate local reactor into ph-cli ([92f2f530](https://github.com/powerhouse-inc/powerhouse/commit/92f2f530))
+- **codegen:** regenerate lockfile ([31c2d0c5](https://github.com/powerhouse-inc/powerhouse/commit/31c2d0c5))
+- **ph-cli:** update cli to format by default ([7418e777](https://github.com/powerhouse-inc/powerhouse/commit/7418e777))
+- **codegen:** remove format generated from action ([5d7e1c48](https://github.com/powerhouse-inc/powerhouse/commit/5d7e1c48))
+- **codegen:** use prettier api to format typescript from gql ([8896d86e](https://github.com/powerhouse-inc/powerhouse/commit/8896d86e))
+
+### 🩹 Fixes
+
+- **design-system:** fix scalars import ([#506](https://github.com/powerhouse-inc/powerhouse/pull/506))
+
+### ❤️  Thank You
+
+- acaldas
+- Guillermo Puente Sandoval
+- ryanwolhuter
+
+## 1.8.0 (2024-11-11)
+
+### 🚀 Features
+
+- **scalars, design-system:** added EthereumAddress and AmountTokens scalars ([e5124e19](https://github.com/powerhouse-inc/powerhouse/commit/e5124e19))
+
+### ❤️  Thank You
+
+- Guillermo Puente Sandoval
+
+## 1.7.0 (2024-11-06)
+
+### 🚀 Features
+
+- **document-model-libs:** regenerate lockfile ([0effee2e](https://github.com/powerhouse-inc/powerhouse/commit/0effee2e))
+- qa updates 4 ([50ac4eb5](https://github.com/powerhouse-inc/powerhouse/commit/50ac4eb5))
+- **document-model-libs:** add more inputs ([37db4561](https://github.com/powerhouse-inc/powerhouse/commit/37db4561))
+
+### ❤️  Thank You
+
+- ryanwolhuter @ryanwolhuter
+
+## 1.6.0 (2024-11-05)
+
+### 🚀 Features
+
+- **document-model-libs:** regenerate lockfile ([cf76e91d](https://github.com/powerhouse-inc/powerhouse/commit/cf76e91d))
+- **document-model-libs:** address initial QA feedback ([7513155e](https://github.com/powerhouse-inc/powerhouse/commit/7513155e))
+
+### ❤️  Thank You
+
+- ryanwolhuter @ryanwolhuter
+
+## 1.5.1 (2024-11-01)
+
+### 🚀 Features
+
+- **codegen:** changed bundle to esm ([24b33b50](https://github.com/powerhouse-inc/powerhouse/commit/24b33b50))
+- **switchboard-gui:** updated gitignore ([1fd60a3b](https://github.com/powerhouse-inc/powerhouse/commit/1fd60a3b))
+- **reactor-api:** init project ([#388](https://github.com/powerhouse-inc/powerhouse/pull/388))
+- **document-model-libs:** bump deps ([68a98ee6](https://github.com/powerhouse-inc/powerhouse/commit/68a98ee6))
+- **document-model-libs:** re-add prevent default and handle operation initial schema ([ae010779](https://github.com/powerhouse-inc/powerhouse/commit/ae010779))
+- **document-model-libs:** add json editors with sync ([76acd807](https://github.com/powerhouse-inc/powerhouse/commit/76acd807))
+- **document-model-libs:** include typedefs in initial hidden schema state ([88e184fd](https://github.com/powerhouse-inc/powerhouse/commit/88e184fd))
+
+### 🩹 Fixes
+
+- **ph-cli:** bundle cli as esm package and added missing dependencies ([a5a665ef](https://github.com/powerhouse-inc/powerhouse/commit/a5a665ef))
+- **codegen:** fixed bad import ([eefb7b2f](https://github.com/powerhouse-inc/powerhouse/commit/eefb7b2f))
+- bad package reference was breaking build -- also a readme update ([4121e51b](https://github.com/powerhouse-inc/powerhouse/commit/4121e51b))
+- **switchboard:** startup ([#469](https://github.com/powerhouse-inc/powerhouse/pull/469))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+- Benjamin Jordan
+- frankp.eth @froid1911
+- Guillermo Puente @gpuente
+- ryanwolhuter @ryanwolhuter
+
+## 1.5.0 (2024-10-28)
+
+### 🚀 Features
+
+- **scalars:** added default scalars ([#444](https://github.com/powerhouse-inc/powerhouse/pull/444))
+- **document-model-libs:** add initial state editor ([#443](https://github.com/powerhouse-inc/powerhouse/pull/443))
+- **codegen:** integrate scalars into code generation ([#436](https://github.com/powerhouse-inc/powerhouse/pull/436))
+
+### ❤️  Thank You
+
+- Guillermo Puente Sandoval @gpuente
+- Ryan Wolhuter @ryanwolhuter
+
+## 1.4.0 (2024-10-25)
+
+### 🚀 Features
+
+- **document-model-libs:** added scalars as a peerDep ([#435](https://github.com/powerhouse-inc/powerhouse/pull/435))
+- **scalars:** added exports for codegen ([#434](https://github.com/powerhouse-inc/powerhouse/pull/434))
+
+### ❤️  Thank You
+
+- Guillermo Puente Sandoval @gpuente
+
+## 1.3.0 (2024-10-25)
+
+### 🚀 Features
+
+- **scalars:** added amount percentage scalar ([#433](https://github.com/powerhouse-inc/powerhouse/pull/433))
+- **document-model-libs:** simplify editor ([1725f876](https://github.com/powerhouse-inc/powerhouse/commit/1725f876))
+
+### ❤️  Thank You
+
+- Guillermo Puente Sandoval @gpuente
+- ryanwolhuter @ryanwolhuter
+
+## 1.2.0 (2024-10-22)
+
+### 🚀 Features
+
+- **document-model-libs:** fix lint error ([37d4061d](https://github.com/powerhouse-inc/powerhouse/commit/37d4061d))
+- **document-model-libs:** port code from private repo ([11ef336a](https://github.com/powerhouse-inc/powerhouse/commit/11ef336a))
+- **ph-cli:** Support for dev and generate command ([#406](https://github.com/powerhouse-inc/powerhouse/pull/406))
+- **ph-cli:** added base ph-cli package setup ([#404](https://github.com/powerhouse-inc/powerhouse/pull/404))
+- **switchboard-gui:** init ([#405](https://github.com/powerhouse-inc/powerhouse/pull/405))
+- **codegen:** bundle with tsup ([3cccbdf9](https://github.com/powerhouse-inc/powerhouse/commit/3cccbdf9))
+- **document-model-libs:** update codegen dependency ([f5330139](https://github.com/powerhouse-inc/powerhouse/commit/f5330139))
+
+### 🩹 Fixes
+
+- **document-model-libs:** added type annotation for createDocumentStory ([6c5441f3](https://github.com/powerhouse-inc/powerhouse/commit/6c5441f3))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+- frankp.eth @froid1911
+- Guillermo Puente Sandoval @gpuente
+- ryanwolhuter @ryanwolhuter
+
+## 1.1.0 (2024-10-14)
+
+### 🚀 Features
+
+- added codegen build step ([66523278](https://github.com/powerhouse-inc/powerhouse/commit/66523278))
+
+### ❤️  Thank You
+
+- acaldas @acaldas
+
 # 1.0.0 (2024-10-11)
 
 ### 🚀 Features